+++ clinica/src/main/java/com/proyecto/clinica/service/TurnoService.java
@@ -1,41 +1,73 @@
 package com.proyecto.clinica.service;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.proyecto.clinica.entities.Paciente;
+import com.proyecto.clinica.entities.PacienteDTO;
 import com.proyecto.clinica.entities.Turno;
+import com.proyecto.clinica.entities.TurnoDTO;
+import com.proyecto.clinica.exceptions.ResourceNotFoundException;
 import com.proyecto.clinica.repository.impl.TurnoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 @Service
-public class TurnoService {
-
-    private final TurnoRepository turnoRepository;
+public class TurnoService implements ITurnoService{
 
     @Autowired
-    public TurnoService(TurnoRepository turnoRepository) {
-        this.turnoRepository = turnoRepository;
+    private TurnoRepository turnoRepository;
+
+    @Autowired
+    ObjectMapper objectMapper;
+
+    public void guardarTurno(TurnoDTO turnoDTO) {
+        Turno turno = objectMapper.convertValue(turnoDTO,Turno.class);
+        turnoRepository.save(turno);
     }
 
-    public Turno guardar(Turno t) {
-        //t.setFechaIngreso(new Date());
-        return turnoRepository.save(t);
+
+    @Override
+    public void crearTurno(TurnoDTO turnoDTO) {
+        guardarTurno(turnoDTO);
     }
 
-    public void eliminar(Long id) {
+    @Override
+    public TurnoDTO leerTurno(Long id) throws ResourceNotFoundException {
+        Optional<Turno> turno = turnoRepository.findById(id);
+        TurnoDTO turnoDTO = null;
+        if(turno.isPresent()){
+            turnoDTO=objectMapper.convertValue(turno, TurnoDTO.class);
+        }else{
+            throw new ResourceNotFoundException("No existe un turno con el id " + id);
+        }
+        return turnoDTO;
+    }
+
+    @Override
+    public void modificarTurno(TurnoDTO turnoDTO) {
+        guardarTurno(turnoDTO);
+    }
+
+    @Override
+    public void eliminarTurno(Long id) throws ResourceNotFoundException {
+        if (leerTurno(id)==null){
+            throw new ResourceNotFoundException("No existe un turno con el id " + id);
+        }
+
         turnoRepository.deleteById(id);
     }
 
-    public Optional<Turno>/*Turno*/ buscar(Long id) {
-        return turnoRepository.findById(id);
-    }
-
-    public List<Turno> listar() {
-        return turnoRepository.findAll();
-    }
-
-    public Turno actualizar(Turno t) {
-        return turnoRepository.save(t);
+    @Override
+    public Set<TurnoDTO> listarTodos() {
+        List<Turno> turnos = turnoRepository.findAll();
+        Set<TurnoDTO> turnoDTOS = new HashSet<>();
+        for (Turno turno: turnos){
+            turnoDTOS.add(objectMapper.convertValue(turno, TurnoDTO.class));
+        }
+        return turnoDTOS;
     }
 }
